Adopt new API and other changes
Adopt the new PurpleAir api, along with a number of other changes
I made in the last year or so.

Add a check for freshness of data, as a friend who uses this
had his sensor go offline for a month and didn't know. If the data
is more than an hour old, the AQI value is bracketed by two
asterisks.

Changed to AM/PM time.